% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{choose_dict}
\alias{choose_dict}
\title{Select the appropriate dictionary - with or without stresses?}
\usage{
choose_dict(keep_stresses)
}
\arguments{
\item{keep_stresses}{keep the stresses attached to each phonmeme?}
}
\description{
Select the appropriate dictionary - with or without stresses?
}
