% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmudict.R
\docType{data}
\name{cmudict}
\alias{cmudict}
\title{CMU Pronouncing Dictionary.}
\format{An object of class \code{character} of length 133854.}
\usage{
cmudict
}
\description{
The Carnegie Mellon University Pronouncing Dictionary is an open-source
machine-readable pronunciation dictionary for North American English that
contains over 134,000 words and their pronunciations.
}
\details{
This is a named vector of words and their pronunciations.

The CMU Pronouncing Dictionary is Copyright (C) 1993-2015 Carnegie Mellon University.

The full copyright notice is available as part of this package and may be
found by calling: \code{system.file('COPYRIGHT-CMU-Pronouncing-Dictionary.txt',
package = 'phon')}
}
\keyword{datasets}
