% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contains_phonemes.R
\name{contains_phonemes}
\alias{contains_phonemes}
\title{Find words which contain the given phoneme sequence}
\usage{
contains_phonemes(phonemes, keep_stresses = FALSE)
}
\arguments{
\item{phonemes}{a character string of phonemes to look for}

\item{keep_stresses}{keep the stresses attached to each phonmeme? default: FALSE
If FALSE, more likely to find matches, but may be less accurate.}
}
\value{
Character vector of words which contain the given phonemes
}
\description{
Find words which contain the given phoneme sequence
}
\examples{
# Find all words which contain the sound of "through" when spoken
phons <- phonemes("through")[[1]]
contains_phonemes(phons)

}
