% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contains-pronunciation.R
\name{contains_pronunciation}
\alias{contains_pronunciation}
\title{Find words which contain pronunciation of the given word}
\usage{
contains_pronunciation(word, keep_stresses = FALSE)
}
\arguments{
\item{word}{single word}

\item{keep_stresses}{keep the stresses attached to each phonmeme? Default: FALSE}
}
\value{
Character vector of words which contain the pronunciation of the
        given word.
}
\description{
Find words which contain pronunciation of the given word
}
\examples{
# Find all words which contain the sound of "through" when spoken
contains_pronunciation('through')

}
