% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contains-pronunciation.R
\name{contains_pronunciation_phonemes}
\alias{contains_pronunciation_phonemes}
\title{Find words which contain the given phoneme character strings}
\usage{
contains_pronunciation_phonemes(phons, keep_stresses = FALSE)
}
\arguments{
\item{phons}{Vector of phoneme strings e.g. c("K AE1 R IY0", "K EH1 R IY0")}

\item{keep_stresses}{keep the stresses attached to each phonmeme? Default: FALSE}
}
\value{
Character vector of words which contain the given phonemes
}
\description{
Find words which contain the given phoneme character strings
}
\examples{
# Find all words which contain the sound of "through" when spoken
phons <- phonemes("through")
contains_pronunciation_phonemes(phons)

}
