% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{create_cmudict_as_data_frame}
\alias{create_cmudict_as_data_frame}
\title{Export the phonemes data as a data.frame}
\usage{
create_cmudict_as_data_frame()
}
\value{
phonemes data as a data.frame of words and phonemes. The phonemes
for each word are collapsed into a single string, and included in both
the original format, along with a version with the stress indicators removed.
}
\description{
The actual \code{cmudict} data is stored within \code{phon} in a
number of pre-processed forms. This reduces overall memory footprint,
and enables faster lookups of information about each word.
}
\details{
\code{create_cmudict_as_data_frame()} creates a more manageable representation of the
\code{cmudict} for external use.
}
\examples{
create_cmudict_as_data_frame()

}
