% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{create_cmudict_as_list}
\alias{create_cmudict_as_list}
\title{Export the phonemes data as a list}
\usage{
create_cmudict_as_list()
}
\value{
Named list of phoneme vectors (one for each word in the dictionary).
}
\description{
The actual \code{cmudict} data is stored within \code{phon} in a
number of pre-processed forms. This reduces overall memory footprint,
and enables faster lookups of information about each word.
}
\details{
\code{create_cmudict_as_list()} creates a more manageable representation of the
\code{cmudict} for external use.
}
\examples{
create_cmudict_as_list()

}
