% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cmudict.R
\name{get_cmudict}
\alias{get_cmudict}
\title{Get a copy of the cmudict}
\usage{
get_cmudict()
}
\value{
named vector of phonemes
}
\description{
The cmudict used internal to the package is pre-processed for faster lookups
and to save memory.
}
\details{
This function creates a named vector of phonemes for use external to the
package
}
