% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_word_idxs}
\alias{get_word_idxs}
\title{Internal helper function to lookup the indices at which this word occurs}
\usage{
get_word_idxs(word)
}
\arguments{
\item{word}{find all idxs of this word}
}
\value{
integer vector of indices
}
\description{
Words may occur multiple time in the dictionary because of multiple
pronunciations
}
