% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homophones.R
\name{homophones}
\alias{homophones}
\title{Find all homophones of the given word}
\usage{
homophones(word)
}
\arguments{
\item{word}{find homophones for this word}
}
\value{
character vector of homophones.
}
\description{
Homophones are words with the same pronunciation but different spelling.
}
\details{
This function finds the phonemes (without stresses) for the given word, and
finds all other words with the same phonemes.
}
\examples{
homophones("steak")
homophones("carry")

}
