% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homophones.R
\name{homophones_phonemes}
\alias{homophones_phonemes}
\title{Find all homophones of the given phoneme character strings}
\usage{
homophones_phonemes(phons, keep_stresses = FALSE)
}
\arguments{
\item{phons}{Vector of phoneme strings e.g. c("K AE1 R IY0", "K EH1 R IY0")}

\item{keep_stresses}{keep the stresses attached to each phonmeme? Default: FALSE}
}
\value{
character vector of homophones.
}
\description{
Homophones are words with the same pronunciation but different spelling.
}
\details{
This function finds the phonemes (without stresses) for the given word, and
finds all other words with the same phonemes.
}
\examples{
homophones_phonemes(phonemes('steak'))
homophones_phonemes("K AH R IY")

}
