% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phonemes.R
\name{phonemes}
\alias{phonemes}
\title{Return phonemes for the given word.}
\usage{
phonemes(word, keep_stresses = FALSE)
}
\arguments{
\item{word}{single word}

\item{keep_stresses}{keep the stresses attached to each phonmeme? Default: FALSE}
}
\value{
Vector of phoneme strings. Each string is one variation on
        the pronunciation of the given word, represented as the
        ARPABET phonetic transcription.
}
\description{
Return phonemes for the given word.
}
\examples{
phonemes("cellar")

}
