% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_stresses}
\alias{remove_stresses}
\title{Remove stresses from phonemes}
\usage{
remove_stresses(x)
}
\arguments{
\item{x}{phonemes}
}
\value{
modified phonemes with stresses removed
}
\description{
Remove stresses from phonemes
}
