% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhymes.R
\name{rhymes}
\alias{rhymes}
\title{Find rhymes for the given word}
\usage{
rhymes(word, match_length = 2:13)
}
\arguments{
\item{word}{word for which to find rhymes}

\item{match_length}{how good a rhyme? the higher the number the more
trailing phoneme matches between the word and any
selected rhyming word. Use a range. Default: 2:13.
Min = 2, Max = 13}
}
\value{
list of rhymes separated by rhyme intensity level
}
\description{
To find rhymes, \code{phon} compares trailing phonemes i.e. if the phonemes at the end of a
word in the dictionary match those at the end of the given word, then they rhyme.
}
\details{
The rhymes are returned in multiple vectors:
\itemize{
\item{Words with the most matching trailing phonemes are returned first.}
\item{Subsequent vectors have fewer matching trailing phonemes.}
\item{Words in earlier vectors are repeated in later vectors.}
\item{The names of the list are the number of trailing phonemes which match.}
\item{If there are no rhymes of the given match length, then no space in the
      returned list is reserved for this empty vector.}
}
}
\examples{
rhymes("drudgery")

}
