% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhymes.R
\name{rhymes_core}
\alias{rhymes_core}
\title{Find rhymes for a given trailing phoneme match length}
\usage{
rhymes_core(match_length, word)
}
\arguments{
\item{match_length}{how good a rhyme? the higher the number the more
trailing phoneme matches between the word and any
selected rhyming word. Use a range. Default: 2:13.
Min = 2, Max = 13}

\item{word}{word for which to find rhymes}
}
\value{
character vector of words (excluding the given word). returns
\code{character(0)} if no rhymes found.
}
\description{
Look up the word in the 'last_syls_group' data
}
