% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhymes.R
\name{rhymes_phonemes_single}
\alias{rhymes_phonemes_single}
\alias{rhymes_phonemes}
\title{Find rhymes for the given phoneme strings}
\usage{
rhymes_phonemes_single(phons, keep_stresses = FALSE, min_phonemes = 3L)

rhymes_phonemes(phons, keep_stresses = FALSE, min_phonemes = 2L)
}
\arguments{
\item{phons}{Vector of phoneme strings e.g. c("K AE1 R IY0", "K EH1 R IY0")}

\item{keep_stresses}{keep the stresses attached to each phonmeme? Default: FALSE}

\item{min_phonemes}{the minimum number of matching phonemes. Default: 3}
}
\value{
data.frame of rhymes with \code{rhyme_length} giving the number of
trailing phonems which match the given word.
}
\description{
To find rhymes, \code{phon} compares trailing phonemes. Words rhyme with each
other if they have the same phonemes at the end.
}
\examples{
rhymes_phonemes_single("D R AH1 JH ER0 IY0")

}
