% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sanitize_phons}
\alias{sanitize_phons}
\title{Sanitize phons}
\usage{
sanitize_phons(phons)
}
\arguments{
\item{phons}{Vector of phoneme character strings.}
}
\value{
Sanitized phons
}
\description{
ensure only a single space between individual phonemes, and that there
are no spaces at either end
}
