% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similar.R
\name{similar}
\alias{similar}
\title{Find words with a similar phonetic makeup}
\usage{
similar(word, phoneme_mismatches = 1L)
}
\arguments{
\item{word}{Find words similar to this}

\item{phoneme_mismatches}{Number of phonemes which can be different between
between words. Default 1}
}
\value{
character vector of similar words
}
\description{
Similar sounding words are found by comparing words with the same number
of phonemes but with a number of mismatches allowed.
}
\examples{
similar("statistics", phoneme_mismatches = 5)

}
