% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similar.R
\name{similar_core}
\alias{similar_core}
\title{Find words with a similar phonetic makeup to the given phonetic encoding}
\usage{
similar_core(this_phon, phoneme_mismatches = 1L)
}
\arguments{
\item{this_phon}{character vector of phonetic encoding}

\item{phoneme_mismatches}{Maximum number of phonemes which can differ. Default 1}
}
\description{
Find words with a similar phonetic makeup to the given phonetic encoding
}
