% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sounds-like.R
\name{sounds_like}
\alias{sounds_like}
\title{Find words with a similar phonemes to the given word}
\usage{
sounds_like(word, keep_stresses = FALSE, phoneme_mismatches = 1L)
}
\arguments{
\item{word}{single word}

\item{keep_stresses}{keep the stresses attached to each phonmeme? Default: FALSE}

\item{phoneme_mismatches}{Maximum number of phonemes which can be different between
between words. Default: 1}
}
\value{
Character vector of similar words.
}
\description{
Similar sounding words are found by comparing words with the same number
of phonemes but with a number of mismatches allowed.
}
\examples{
sounds_like("statistics", phoneme_mismatches = 5)

}
