% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sounds-like.R
\name{sounds_like_phonemes_single}
\alias{sounds_like_phonemes_single}
\title{Find words with a similar phonemes to the given phoneme string}
\usage{
sounds_like_phonemes_single(phons, keep_stresses = FALSE,
  phoneme_mismatches = 1L)
}
\arguments{
\item{phons}{Vector of phoneme strings e.g. c("K AE1 R IY0", "K EH1 R IY0")}

\item{keep_stresses}{keep the stresses attached to each phonmeme? Default: FALSE}

\item{phoneme_mismatches}{Maximum number of phonemes which can differ. Default 1}
}
\description{
Find words with a similar phonemes to the given phoneme string
}
