% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sounds_similar.R
\name{sounds_similar_core}
\alias{sounds_similar_core}
\title{Find words with a similar phonetic makeup to the given phonetic encoding}
\usage{
sounds_similar_core(this_phon, phoneme_mismatches = 1L)
}
\arguments{
\item{this_phon}{character vector of phonetic encoding}

\item{phoneme_mismatches}{Maximum number of phonemes which can differ. Default 1}
}
\description{
Find words with a similar phonetic makeup to the given phonetic encoding
}
