% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syllables.R
\name{syllables}
\alias{syllables}
\title{Number of syllables in a word}
\usage{
syllables(word)
}
\arguments{
\item{word}{count syllables in these word}
}
\value{
Vector of counts of syllables in words. Returns \code{NA_Integer} if word
is not in the CMU Pronouncing Dictionary.
}
\description{
The number of syllables in a word is the count of the number of stressed phonemes
in the word.
}
\examples{
syllables("average")
syllables("antidisestablishmentarianism")

}
