% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syllables.R
\name{syllables_phonemes}
\alias{syllables_phonemes}
\title{Calculate number of syllables in each phoneme string.}
\usage{
syllables_phonemes(phons)
}
\arguments{
\item{phons}{Vector of phoneme strings e.g. c("K AE1 R IY0", "K EH1 R IY0")}
}
\value{
Integer vector of counts of syllables in each phoneme string.
}
\description{
The number of syllables is the count of the number of stressed phonemes
in the word.
}
\examples{
syllables_phonemes(phonemes("average", keep_stresses = TRUE))
syllables_phonemes(c("G R EY1 T", "F AE0 N T AE1 S T IH0 K"))

}
