% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homophones.R
\name{homophones}
\alias{homophones}
\title{Find all homophones of the given word}
\usage{
homophones(word, keep_stresses = FALSE)
}
\arguments{
\item{word}{single word}

\item{keep_stresses}{keep the stresses attached to each phonmeme? Default: FALSE}
}
\value{
Character vector of words which are homophones for the given word.
}
\description{
Homophones are words with the same pronunciation but different spelling.
}
\examples{
homophones("steak")
homophones("carry")

}
