% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phonemes.R
\name{phonemes}
\alias{phonemes}
\title{Return phonemes for the given word}
\usage{
phonemes(word, keep_stresses = FALSE)
}
\arguments{
\item{word}{word to lookup in CMU Pronouncing Dictionary}

\item{keep_stresses}{keep the phoneme stresses? default: FALSE}
}
\value{
Character vector. Each entry is one variation on
        the pronunciation of the given word. The character vectors are the
        ARPABET phonetic transcription codes of the word
}
\description{
This function returns the original phonemes as provided in \code{cmudict}.
}
\examples{
phonemes("cellar")

}
