% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhymes.R
\name{rhymes}
\alias{rhymes}
\title{Find rhymes for the given word}
\usage{
rhymes(word, keep_stresses = FALSE, min_phonemes = 2L)
}
\arguments{
\item{word}{word for which to find rhymes}

\item{keep_stresses}{keep the phoneme stresses? default: FALSE}

\item{min_phonemes}{the minimum number of matching phonemes. Default: 2L}
}
\value{
list of rhymes grouped by number of mathching trailing phonemes with
the supplied word
}
\description{
To find rhymes, \code{phon} compares trailing phonemes i.e. if the phonemes at the end of a
word in the dictionary match those at the end of the given word, then they rhyme.
}
\details{
The rhymes are returned in multiple vectors:
\itemize{
\item{Words with the most matching trailing phonemes are returned first.}
\item{Subsequent vectors have fewer matching trailing phonemes.}
\item{Words in earlier vectors are repeated in later vectors.}
}
}
\examples{
rhymes("drudgery")

}
