% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syllables.R
\name{syllables}
\alias{syllables}
\title{Calculate number of syllables in a word.}
\usage{
syllables(word, all_pronunciations = FALSE)
}
\arguments{
\item{word}{single word}

\item{all_pronunciations}{If TRUE, return the syllable counts for all pronunciations,
otherwise only return syllable count for the first pronunciation.
Default: FALSE}
}
\value{
Integer vector of counts of syllables in word. Returns \code{NA_Integer} if word
is not in the CMU Pronouncing Dictionary.
}
\description{
The number of syllables in a word is the count of the number of stressed phonemes
in the word.
}
\examples{
syllables("average")
syllables("antidisestablishmentarianism")

}
